+++ src/styles/_dashboard.scss
@@ -1,4 +1,4 @@
-// FONTS
+// TODO: decide fonts, and import them in _typography.scss
 // @font-face {
 //   font-family: "Route159";
 //   src: local("Route159"),
