+++ src/styles/_faq.scss
@@ -1,3 +1,4 @@
+// TODO: styling with classes, id for linking or specific stuff
 #faq {
   background: linear-gradient(rgba(0, 0, 0, 0.5), rgba(0, 0, 0, 0.5)),
     url("../assets/intro-background.jpg");
