--- src/styles/_mentorStats.scss
@@ -1,3 +1,0 @@
-.mentor-stats {
-  text-align: "center" !important;
-}
