+++ src/styles/_stats.scss
@@ -1,39 +1,3 @@
-* {
-  text-rendering: optimizeLegibility;
-  box-sizing: border-box;
-}
-
-html {
-  margin: 0;
-}
-
-body {
-  margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue",
-    sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  background-color: $background-color-page;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
-}
-
-.App {
-  .hero {
-    h1 {
-      font-size: 5rem;
-    }
-  }
-}
-
-section {
-  margin-bottom: 5rem;
-}
-
 // the following style code taken from https://codepen.io/colewaldrip/pen/bdZVGd
 .refresh-icon {
   font-size: 1.5rem;
@@ -47,6 +11,7 @@ section {
   transition: color 0.2s ease, background-color 0.2s ease, transform 0.3s ease;
 }
 
+// TODO: nest the following declarations
 .refresh-icon:after {
   content: "";
   width: 100%;
@@ -260,4 +225,8 @@ section {
       opacity: 0.5;
     }
   }
+}
+
+.mentor-stats {
+  text-align: "center" !important;
 }
