+++ src/styles/_vars.scss
@@ -1,3 +1,5 @@
+// TODO: new and more variables, and import them wherever hardcoded
+
 // ref:https://blog.prototypr.io/designing-a-dark-mode-for-your-ios-app-the-ultimate-guide-6b043303b941
 // colors for dark mode
 // background
