+++ src/views/Testimonials.js
@@ -13,6 +13,7 @@ export default function Testimonial() {
           </div>
         </div>
       </section>
+      {/* TODO: maybe a markdown renderer here if time permits, reading and editing these shouldn't require writing code */}
       <section>
         <div className="container">
           <div className="columns is-multiline is-centered">
